+++ src/assets/icons/operat/refresh.ts
@@ -1,0 +1,1 @@
+export const ICON_BODY: string = '<path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M20 11A8.1 8.1 0 0 0 4.5 9M4 5v4h4m-4 4a8.1 8.1 0 0 0 15.5 2m.5 4v-4h-4" />';