+++ src/assets/icons/sign/notEquals.ts
@@ -1,0 +1,1 @@
+export const ICON_BODY: string = '<path fill="none" stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 10h14M5 14h14M5 19L19 5" />';